requirements.txt
requests
apache-airflow==2.5.0  # Ensure you're using the version you need
psycopg2-binary